requirements.txt
absl-py==2.1.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
basicsr==1.4.2
beautifulsoup4==4.12.3
blendmodes==2022
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.4.1
exceptiongroup==1.2.0
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.3.2
filelock==3.13.1
filterpy==1.4.5
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
gdown==5.1.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.32
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.62.0
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imhist==0.0.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
imWatermark==0.0.2
inflection==0.5.1
invisible-watermark==0.2.0
jax==0.4.24
Jinja2==3.1.3
joblib==1.4.2
jsonmerge==1.8.0
jsonpickle==3.2.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.7
lark==1.1.2
lazy_loader==0.3
librosa==0.10.2.post1
lightning-utilities==0.10.1
llvmlite==0.42.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mediapipe==0.10.10
mido==1.3.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
music21==9.1.0
natsort==8.4.0
networkx==3.2.1
numba==0.59.0
numpy==1.23.5
omegaconf==2.2.3
open-clip-torch==2.20.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.14
packaging==23.2
pandas==2.2.0
piexif==1.1.3
Pillow==9.5.0
platformdirs==4.2.0
pooch==1.8.2
pretty-midi==0.2.10
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning==1.9.4
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
realesrgan==0.3.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
resize-right==0.0.2
rich==13.7.0
rpds-py==0.18.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
starlette==0.26.1
streamlit==1.31.1
streamlit-keyup==0.2.3
sympy==1.12
tb-nightly==2.17.0a20240222
tenacity==8.2.3
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.2.12
timm==0.9.2
tokenizers==0.13.3
tomesd==0.1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchdiffeq==0.2.3
torchmetrics==1.3.1
torchsde==0.2.6
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
trampoline==0.1.2
transformers==4.30.2
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.17
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
wandb==0.16.3
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.6.0
websockets==11.0.3
Werkzeug==3.0.1
xformers==0.0.24
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
